moved meta tags to it's own module and config
